requirements.txt
anthropic
ccxt
python-telegram-bot
sqlalchemy
asyncpg
apscheduler
streamlit
httpx
aiohttp
redis
python-dotenv
feedparser
psycopg2-binary
plotly
pandas
